Move SchemaField to public types
