+++ src/sources/types.ts
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-import type {SchemaField} from '../types-internal.js';
-import type {Filters, QueryParameters} from '../types.js';
+import type {Filters, SchemaField, QueryParameters} from '../types.js';
 
 export type SourceRequiredOptions = {
   /** Carto platform access token. */
